+++ jablib/src/test/java/org/jabref/logic/search/query/SearchQueryExtractorConversionTest.java
@@ -1,8 +1,10 @@
 package org.jabref.logic.search.query;
 
+import java.util.EnumSet;
 import java.util.List;
 import java.util.stream.Stream;
 
+import org.jabref.model.search.SearchFlags;
 import org.jabref.model.search.query.SearchQuery;
 import org.jabref.model.search.query.SearchQueryNode;
 
@@ -15,26 +17,31 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 public class SearchQueryExtractorConversionTest {
     public static Stream<Arguments> searchConversion() {
         return Stream.of(
-                Arguments.of(List.of("term"), "term"),
-                Arguments.of(List.of("regex.*term"), "regex.*term"),
-                Arguments.of(List.of("term"), "any = term"),
-                Arguments.of(List.of("term"), "any CONTAINS term"),
-                Arguments.of(List.of("a", "b"), "a AND b"),
-                Arguments.of(List.of("a", "b", "c"), "a OR b AND c"),
-                Arguments.of(List.of("a", "b"), "a OR b AND NOT c"),
-                Arguments.of(List.of("a", "b"), "author = a AND title = b"),
-                Arguments.of(List.of(), "NOT a"),
-                Arguments.of(List.of("a", "b", "c"), "(any = a OR any = b) AND NOT (NOT c AND title = d)"),
-                Arguments.of(List.of("b", "c"), "title != a OR b OR c"),
-                Arguments.of(List.of("a", "b"), "a b"),
-                Arguments.of(List.of("term1 term2"), "\"term1 term2\"")
+                Arguments.of(List.of(), "", EnumSet.noneOf(SearchFlags.class)),
+                Arguments.of(List.of("term"), "term", EnumSet.noneOf(SearchFlags.class)),
+                Arguments.of(List.of("regex\\.\\*term"), "regex.*term", EnumSet.noneOf(SearchFlags.class)),
+                Arguments.of(List.of("regex.*term"), "regex.*term", EnumSet.of(SearchFlags.REGULAR_EXPRESSION)),
+                Arguments.of(List.of("term"), "any = term", EnumSet.noneOf(SearchFlags.class)),
+                Arguments.of(List.of("term"), "any CONTAINS term", EnumSet.noneOf(SearchFlags.class)),
+                Arguments.of(List.of("regex.*term"), "any =~ regex.*term", EnumSet.noneOf(SearchFlags.class)),
+                Arguments.of(List.of("a", "b"), "a AND b", EnumSet.noneOf(SearchFlags.class)),
+                Arguments.of(List.of("a", "b", "c"), "a OR b AND c", EnumSet.noneOf(SearchFlags.class)),
+                Arguments.of(List.of("a", "b"), "a OR b AND NOT c", EnumSet.noneOf(SearchFlags.class)),
+                Arguments.of(List.of("a", "b"), "author = a AND title = b", EnumSet.noneOf(SearchFlags.class)),
+                Arguments.of(List.of(), "NOT a", EnumSet.noneOf(SearchFlags.class)),
+                Arguments.of(List.of("a", "b", "c"), "(any = a OR any = b) AND NOT (NOT c AND title = d)", EnumSet.noneOf(SearchFlags.class)),
+                Arguments.of(List.of("b", "c"), "title != a OR b OR c", EnumSet.noneOf(SearchFlags.class)),
+                Arguments.of(List.of("a", "b"), "a b", EnumSet.noneOf(SearchFlags.class)),
+                Arguments.of(List.of("term1 term2"), "\"term1 term2\"", EnumSet.noneOf(SearchFlags.class)),
+                Arguments.of(List.of("t\\(1\\)erm"), "t\\(1\\)erm", EnumSet.noneOf(SearchFlags.class)),
+                Arguments.of(List.of("C:\\\\temp"), "\"C:\\temp\"", EnumSet.noneOf(SearchFlags.class))
         );
     }
 
     @ParameterizedTest
     @MethodSource
-    void searchConversion(List<String> expected, String searchExpression) {
-        List<String> result = SearchQueryConversion.extractSearchTerms(new SearchQuery(searchExpression)).stream().map(SearchQueryNode::term).toList();
+    void searchConversion(List<String> expected, String searchExpression, EnumSet<SearchFlags> searchFlags) {
+        List<String> result = SearchQueryConversion.extractSearchTerms(new SearchQuery(searchExpression, searchFlags)).stream().map(SearchQueryNode::term).toList();
         assertEquals(expected, result);
     }
 }